requirements.txt
Flask
python-dotenv
azure-ai-formrecognizer
